Move to components directory, rename, & format
