+++ styles/ScoredSentenceList.scss
